// src/lib/api/generated/model/v1SkillFile.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * aisphere/options/v1/authz.proto
 * OpenAPI spec version: version not set
 */

export interface V1SkillFile {
  id?: string;
  skillName?: string;
  version?: string;
  path?: string;
  name?: string;
  type?: string;
  size?: string;
  binary?: boolean;
  content?: string;
  createTime?: string;
  updateTime?: string;
}