// src/lib/api/generated/model/v1SkillVersion.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * aisphere/options/v1/authz.proto
 * OpenAPI spec version: version not set
 */

export interface V1SkillVersion {
  id?: string;
  skillName?: string;
  version?: string;
  status?: string;
  author?: string;
  commitMsg?: string;
  publishPipelineInfo?: string;
  downloadCount?: string;
  md5?: string;
  sha256?: string;
  revision?: string;
  sizeBytes?: string;
  manifestJson?: string;
  createTime?: string;
  updateTime?: string;
}